fix deprecation messages for tests and metrics: move some dashboard methods to utils
