+++ dev_requirements.txt
@@ -3,7 +3,7 @@ wheel==0.35.1
 setuptools==50.3.2
 flake8==4.0.1
 jupyter==1.0.0
-mypy==0.910
+mypy==0.920
 pytest==6.2.5
 types-PyYAML==6.0.1
 types-requests==2.26.0
