+++ src/evidently/renderers/notebook_utils.py
@@ -1,5 +1,5 @@
-from evidently.dashboard.dashboard import file_html_template
-from evidently.dashboard.dashboard import inline_template
+from evidently.utils.dashboard import file_html_template
+from evidently.utils.dashboard import inline_template
 
 
 def determine_template(mode: str):
